moving reducer
