+++ src/contexts/responses.tsx
@@ -1,10 +1,46 @@
 import React, { createContext, useReducer } from 'react';
-import Reducer from '../reducers/responsesReducer';
 import { TResponses } from '../types/types';
 
-const testQuestionResponses: TResponses = {
+// -- Reducer ---//
+type Action = {
+  type: 'ADD_SETONE';
+  action: {
+    questionId: number;
+    answerId: number;
+  };
+};
+
+// TO DO - The Assumption at present is that each time we are supplied a question, in it will have not yet been answered. Need to add the functionality to check if the question has already been answered and update if if has.
+
+export function responsesReducer(state: TResponses, action: Action) {
+  switch (action.type) {
+    case 'ADD_SETONE':
+      const newState = {
+        ...state,
+      };
+      newState.SetOne.push({
+        questionId: action.action.questionId,
+        answerId: action.action.answerId,
+      });
+      return newState;
+    default:
+      return state;
+  }
+}
+
+// --- use Responses Hook ---//
+
+const intialResponses: TResponses = {
   SetOne: [],
 };
+
+const useResponses = () => {
+  const [state, dispatch] = useReducer(responsesReducer, intialResponses);
+
+  return { state, dispatch };
+};
+
+// -- Context Provider ---//
 
 export const ResponsesContext = createContext<TResponses>({} as TResponses);
 export const ResponsesReducerContext = createContext<React.Dispatch<any>>(
@@ -12,7 +48,7 @@ export const ResponsesReducerContext = createContext<React.Dispatch<any>>(
 );
 
 export const ResponsesProvider: React.FC = ({ children }) => {
-  const [state, dispatch] = useReducer(Reducer, testQuestionResponses);
+  const { state, dispatch } = useResponses();
 
   return (
     <ResponsesContext.Provider value={state}>
